requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
assemblyai==0.25.0
certifi==2024.2.2
distro==1.9.0
Django==5.0.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.17.1
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
pytube==15.0.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
websockets==12.0
